requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.32.0
unstructured==0.9.2
tiktoken==0.6.0
faiss-cpu==1.8.0
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
OpenAI